Apply calculate performance metrics for model evaluation

Apply calculate performance metrics for binary classification model evaluation.

run_performance(model, actual = NULL)

Arguments

model

A model_df. results of predicted model that created by run_predict().

actual

factor. A data of target variable to evaluate the model. It supports factor that has binary class.

Value

model_df. results of predicted model. model_df is composed of tbl_df and contains the following variables.:

step : character. The current stage in the model fit process. The result of calling run_performance() is returned as "3.Performanced".

model_id : character. Type of fit model.

target : character. Name of target variable.

positive : character. Level of positive class of binary classification.

fitted_model : list. Fitted model object.

predicted : list. Predicted value by individual model. Each value has a predict_class class object.

performance : list. Calculate metrics by individual model. Each value has a numeric vector.

The performance metrics calculated are as follows.:

ZeroOneLoss : Normalized Zero-One Loss(Classification Error Loss).

Accuracy : Accuracy.

Precision : Precision.

Recall : Recall.

Sensitivity : Sensitivity.

Specificity : Specificity.

F1_Score : F1 Score.

Fbeta_Score : F-Beta Score.

LogLoss : Log loss / Cross-Entropy Loss.

AUC : Area Under the Receiver Operating Characteristic Curve (ROC AUC).

Gini : Gini Coefficient.

PRAUC : Area Under the Precision-Recall Curve (PR AUC).

LiftAUC : Area Under the Lift Chart.

GainAUC : Area Under the Gain Chart.

KS_Stat : Kolmogorov-Smirnov Statistic.

Details

run_performance() is performed in parallel when calculating the performance evaluation index. However, it is not supported in MS-Windows operating system and RStudio environment.

Examples

# \donttest{
library(dplyr)

# Divide the train data set and the test data set.
sb <- rpart::kyphosis %>%
  split_by(Kyphosis)

# Extract the train data set from original data set.
train <- sb %>%
  extract_set(set = "train")

# Extract the test data set from original data set.
test <- sb %>%
  extract_set(set = "test")

# Sampling for unbalanced data set using SMOTE(synthetic minority over-sampling technique).
train <- sb %>%
  sampling_target(seed = 1234L, method = "ubSMOTE")

# Cleaning the set.
train <- train %>%
  cleanse
#> ── Checking unique value ─────────────────────────── unique value is one ──
#> No variables that unique value is one.
#> 
#> ── Checking unique rate ─────────────────────────────── high unique rate ──
#> No variables that high unique rate.
#> 
#> ── Checking character variables ─────────────────────── categorical data ──
#> No character variables.
#> 
#> 

# Run the model fitting.
result <- run_models(.data = train, target = "Kyphosis", positive = "present")
result
#> # A tibble: 7 × 7
#>   step     model_id     target   is_factor positive negative fitted_model
#>   <chr>    <chr>        <chr>    <lgl>     <chr>    <chr>    <list>      
#> 1 1.Fitted logistic     Kyphosis TRUE      present  absent   <glm>       
#> 2 1.Fitted rpart        Kyphosis TRUE      present  absent   <rpart>     
#> 3 1.Fitted ctree        Kyphosis TRUE      present  absent   <BinaryTr>  
#> 4 1.Fitted randomForest Kyphosis TRUE      present  absent   <rndmFrs.>  
#> 5 1.Fitted ranger       Kyphosis TRUE      present  absent   <ranger>    
#> 6 1.Fitted xgboost      Kyphosis TRUE      present  absent   <xgb.Bstr>  
#> 7 1.Fitted lasso        Kyphosis TRUE      present  absent   <lognet>    

# Predict the model. (Case 1)
pred <- run_predict(result, test)
pred
#> # A tibble: 7 × 8
#>   step        model_id target is_factor positive negative fitted_model predicted
#>   <chr>       <chr>    <chr>  <lgl>     <chr>    <chr>    <list>       <list>   
#> 1 2.Predicted logistic Kypho… TRUE      present  absent   <glm>        <fct>    
#> 2 2.Predicted rpart    Kypho… TRUE      present  absent   <rpart>      <fct>    
#> 3 2.Predicted ctree    Kypho… TRUE      present  absent   <BinaryTr>   <fct>    
#> 4 2.Predicted randomF… Kypho… TRUE      present  absent   <rndmFrs.>   <fct>    
#> 5 2.Predicted ranger   Kypho… TRUE      present  absent   <ranger>     <fct>    
#> 6 2.Predicted xgboost  Kypho… TRUE      present  absent   <xgb.Bstr>   <fct>    
#> 7 2.Predicted lasso    Kypho… TRUE      present  absent   <lognet>     <fct>    

# Calculate performace metrics. (Case 1)
perf <- run_performance(pred)
perf
#> # A tibble: 7 × 7
#>   step           model_id     target positive fitted_model predicted performance
#>   <chr>          <chr>        <chr>  <chr>    <list>       <list>    <list>     
#> 1 3.Performanced logistic     Kypho… present  <glm>        <fct>     <dbl [15]> 
#> 2 3.Performanced rpart        Kypho… present  <rpart>      <fct>     <dbl [15]> 
#> 3 3.Performanced ctree        Kypho… present  <BinaryTr>   <fct>     <dbl [15]> 
#> 4 3.Performanced randomForest Kypho… present  <rndmFrs.>   <fct>     <dbl [15]> 
#> 5 3.Performanced ranger       Kypho… present  <ranger>     <fct>     <dbl [15]> 
#> 6 3.Performanced xgboost      Kypho… present  <xgb.Bstr>   <fct>     <dbl [15]> 
#> 7 3.Performanced lasso        Kypho… present  <lognet>     <fct>     <dbl [15]> 
perf$performance
#> [[1]]
#> ZeroOneLoss    Accuracy   Precision      Recall Sensitivity Specificity 
#>  0.08333333  0.91666667  1.00000000  0.60000000  0.60000000  1.00000000 
#>    F1_Score Fbeta_Score     LogLoss         AUC        Gini       PRAUC 
#>  0.75000000  0.75000000  0.22463025  0.93684211  0.87368421  0.68545455 
#>     LiftAUC     GainAUC     KS_Stat 
#>  2.20266453  0.84583333 80.00000000 
#> 
#> [[2]]
#> ZeroOneLoss    Accuracy   Precision      Recall Sensitivity Specificity 
#>  0.08333333  0.91666667  1.00000000  0.60000000  0.60000000  1.00000000 
#>    F1_Score Fbeta_Score     LogLoss         AUC        Gini       PRAUC 
#>  0.75000000  0.75000000  0.34208129  0.80000000  0.64210526  0.24166667 
#>     LiftAUC     GainAUC     KS_Stat 
#>  2.53750000  0.73750000 60.00000000 
#> 
#> [[3]]
#> ZeroOneLoss    Accuracy   Precision      Recall Sensitivity Specificity 
#>   0.1666667   0.8333333   0.5714286   0.8000000   0.8000000   0.8421053 
#>    F1_Score Fbeta_Score     LogLoss         AUC        Gini       PRAUC 
#>   0.6666667   0.6666667   0.3977583   0.7894737   0.6210526   0.2722619 
#>     LiftAUC     GainAUC     KS_Stat 
#>   1.5198810   0.7291667  64.2105263 
#> 
#> [[4]]
#> ZeroOneLoss    Accuracy   Precision      Recall Sensitivity Specificity 
#>  0.08333333  0.91666667  0.80000000  0.80000000  0.80000000  0.94736842 
#>    F1_Score Fbeta_Score     LogLoss         AUC        Gini       PRAUC 
#>  0.80000000  0.80000000  0.22061290  0.98947368  0.97894737  0.76333333 
#>     LiftAUC     GainAUC     KS_Stat 
#>  2.33187386  0.88750000 94.73684211 
#> 
#> [[5]]
#> ZeroOneLoss    Accuracy   Precision      Recall Sensitivity Specificity 
#>  0.04166667  0.95833333  1.00000000  0.80000000  0.80000000  1.00000000 
#>    F1_Score Fbeta_Score     LogLoss         AUC        Gini       PRAUC 
#>  0.88888889  0.88888889  0.22338605  0.98947368  0.97894737  0.76333333 
#>     LiftAUC     GainAUC     KS_Stat 
#>  2.33179151  0.88750000 94.73684211 
#> 
#> [[6]]
#> ZeroOneLoss    Accuracy   Precision      Recall Sensitivity Specificity 
#>  0.08333333  0.91666667  0.80000000  0.80000000  0.80000000  0.94736842 
#>    F1_Score Fbeta_Score     LogLoss         AUC        Gini       PRAUC 
#>  0.80000000  0.80000000  0.20825183  0.96842105  0.93684211  0.71964286 
#>     LiftAUC     GainAUC     KS_Stat 
#>  2.27712856  0.87083333 84.21052632 
#> 
#> [[7]]
#> ZeroOneLoss    Accuracy   Precision      Recall Sensitivity Specificity 
#>  0.08333333  0.91666667  1.00000000  0.60000000  0.60000000  1.00000000 
#>    F1_Score Fbeta_Score     LogLoss         AUC        Gini       PRAUC 
#>  0.75000000  0.75000000  0.22545040  0.93684211  0.87368421  0.68545455 
#>     LiftAUC     GainAUC     KS_Stat 
#>  2.20266453  0.84583333 80.00000000 
#> 

# Predict the model. (Case 2)
pred <- run_predict(result, test[, -1])
pred
#> # A tibble: 7 × 8
#>   step        model_id target is_factor positive negative fitted_model predicted
#>   <chr>       <chr>    <chr>  <lgl>     <chr>    <chr>    <list>       <list>   
#> 1 2.Predicted logistic Kypho… TRUE      present  absent   <glm>        <fct>    
#> 2 2.Predicted rpart    Kypho… TRUE      present  absent   <rpart>      <fct>    
#> 3 2.Predicted ctree    Kypho… TRUE      present  absent   <BinaryTr>   <fct>    
#> 4 2.Predicted randomF… Kypho… TRUE      present  absent   <rndmFrs.>   <fct>    
#> 5 2.Predicted ranger   Kypho… TRUE      present  absent   <ranger>     <fct>    
#> 6 2.Predicted xgboost  Kypho… TRUE      present  absent   <xgb.Bstr>   <fct>    
#> 7 2.Predicted lasso    Kypho… TRUE      present  absent   <lognet>     <fct>    

# Calculate performace metrics. (Case 2)
perf <- run_performance(pred, pull(test[, 1]))
perf
#> # A tibble: 7 × 7
#>   step           model_id     target positive fitted_model predicted performance
#>   <chr>          <chr>        <chr>  <chr>    <list>       <list>    <list>     
#> 1 3.Performanced logistic     Kypho… present  <glm>        <fct>     <dbl [15]> 
#> 2 3.Performanced rpart        Kypho… present  <rpart>      <fct>     <dbl [15]> 
#> 3 3.Performanced ctree        Kypho… present  <BinaryTr>   <fct>     <dbl [15]> 
#> 4 3.Performanced randomForest Kypho… present  <rndmFrs.>   <fct>     <dbl [15]> 
#> 5 3.Performanced ranger       Kypho… present  <ranger>     <fct>     <dbl [15]> 
#> 6 3.Performanced xgboost      Kypho… present  <xgb.Bstr>   <fct>     <dbl [15]> 
#> 7 3.Performanced lasso        Kypho… present  <lognet>     <fct>     <dbl [15]> 
perf$performance
#> [[1]]
#> ZeroOneLoss    Accuracy   Precision      Recall Sensitivity Specificity 
#>  0.25000000  0.75000000  0.33333333  0.20000000  0.20000000  0.89473684 
#>    F1_Score Fbeta_Score     LogLoss         AUC        Gini       PRAUC 
#>  0.25000000  0.25000000  0.76115737  0.48421053 -0.03157895  0.17713207 
#>     LiftAUC     GainAUC     KS_Stat 
#>  1.15058165  0.48750000 24.21052632 
#> 
#> [[2]]
#> ZeroOneLoss    Accuracy   Precision      Recall Sensitivity Specificity 
#>  0.25000000  0.75000000  0.33333333  0.20000000  0.20000000  0.89473684 
#>    F1_Score Fbeta_Score     LogLoss         AUC        Gini       PRAUC 
#>  0.25000000  0.25000000  0.57265552  0.54736842  0.09473684  0.21666667 
#>     LiftAUC     GainAUC     KS_Stat 
#>  1.13750000  0.53750000  9.47368421 
#> 
#> [[3]]
#> ZeroOneLoss    Accuracy   Precision      Recall Sensitivity Specificity 
#>   0.4166667   0.5833333   0.1428571   0.2000000   0.2000000   0.6842105 
#>    F1_Score Fbeta_Score     LogLoss         AUC        Gini       PRAUC 
#>   0.1666667   0.1666667   0.7388683   0.4526316  -0.1578947   0.1404762 
#>     LiftAUC     GainAUC     KS_Stat 
#>   0.6655952   0.4625000   0.0000000 
#> 
#> [[4]]
#> ZeroOneLoss    Accuracy   Precision      Recall Sensitivity Specificity 
#>   0.3333333   0.6666667   0.2000000   0.2000000   0.2000000   0.7894737 
#>    F1_Score Fbeta_Score     LogLoss         AUC        Gini       PRAUC 
#>   0.2000000   0.2000000   2.1498921   0.3842105  -0.2210526   0.1307093 
#>     LiftAUC     GainAUC     KS_Stat 
#>   0.9228299   0.4083333  20.0000000 
#> 
#> [[5]]
#> ZeroOneLoss    Accuracy   Precision      Recall Sensitivity Specificity 
#>   0.2916667   0.7083333   0.2500000   0.2000000   0.2000000   0.8421053 
#>    F1_Score Fbeta_Score     LogLoss         AUC        Gini       PRAUC 
#>   0.2222222   0.2222222   0.8046467   0.3684211  -0.2631579   0.1290965 
#>     LiftAUC     GainAUC     KS_Stat 
#>   0.9139033   0.3958333  20.0000000 
#> 
#> [[6]]
#> ZeroOneLoss    Accuracy   Precision      Recall Sensitivity Specificity 
#>   0.3333333   0.6666667   0.2000000   0.2000000   0.2000000   0.7894737 
#>    F1_Score Fbeta_Score     LogLoss         AUC        Gini       PRAUC 
#>   0.2000000   0.2000000   0.8445633   0.4263158  -0.1789474   0.1503852 
#>     LiftAUC     GainAUC     KS_Stat 
#>   1.0454257   0.4416667  20.0000000 
#> 
#> [[7]]
#> ZeroOneLoss    Accuracy   Precision      Recall Sensitivity Specificity 
#>  0.25000000  0.75000000  0.33333333  0.20000000  0.20000000  0.89473684 
#>    F1_Score Fbeta_Score     LogLoss         AUC        Gini       PRAUC 
#>  0.25000000  0.25000000  0.75429467  0.48421053 -0.03157895  0.17713207 
#>     LiftAUC     GainAUC     KS_Stat 
#>  1.15058165  0.48750000 24.21052632 
#> 

# Convert to matrix for compare performace.
sapply(perf$performance, "c")
#>                    [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#> ZeroOneLoss  0.25000000 0.25000000  0.4166667  0.3333333  0.2916667  0.3333333
#> Accuracy     0.75000000 0.75000000  0.5833333  0.6666667  0.7083333  0.6666667
#> Precision    0.33333333 0.33333333  0.1428571  0.2000000  0.2500000  0.2000000
#> Recall       0.20000000 0.20000000  0.2000000  0.2000000  0.2000000  0.2000000
#> Sensitivity  0.20000000 0.20000000  0.2000000  0.2000000  0.2000000  0.2000000
#> Specificity  0.89473684 0.89473684  0.6842105  0.7894737  0.8421053  0.7894737
#> F1_Score     0.25000000 0.25000000  0.1666667  0.2000000  0.2222222  0.2000000
#> Fbeta_Score  0.25000000 0.25000000  0.1666667  0.2000000  0.2222222  0.2000000
#> LogLoss      0.76115737 0.57265552  0.7388683  2.1498921  0.8046467  0.8445633
#> AUC          0.48421053 0.54736842  0.4526316  0.3842105  0.3684211  0.4263158
#> Gini        -0.03157895 0.09473684 -0.1578947 -0.2210526 -0.2631579 -0.1789474
#> PRAUC        0.17713207 0.21666667  0.1404762  0.1307093  0.1290965  0.1503852
#> LiftAUC      1.15058165 1.13750000  0.6655952  0.9228299  0.9139033  1.0454257
#> GainAUC      0.48750000 0.53750000  0.4625000  0.4083333  0.3958333  0.4416667
#> KS_Stat     24.21052632 9.47368421  0.0000000 20.0000000 20.0000000 20.0000000
#>                    [,7]
#> ZeroOneLoss  0.25000000
#> Accuracy     0.75000000
#> Precision    0.33333333
#> Recall       0.20000000
#> Sensitivity  0.20000000
#> Specificity  0.89473684
#> F1_Score     0.25000000
#> Fbeta_Score  0.25000000
#> LogLoss      0.75429467
#> AUC          0.48421053
#> Gini        -0.03157895
#> PRAUC        0.17713207
#> LiftAUC      1.15058165
#> GainAUC      0.48750000
#> KS_Stat     24.21052632
# }