Predict binary classification model

Predict some representative binary classification models.

run_predict(model, .data, cutoff = 0.5)

Arguments

model
A model_df. results of fitted model that created by run_models().

.data
A tbl_df. The data set to predict the model. It also supports tbl, and data.frame objects.

cutoff
numeric. Cut-off that determines the positive from the probability of predicting the positive.

Value

model_df. results of predicted model. model_df is composed of tbl_df and contains the following variables.:

step : character. The current stage in the model fit process. The result of calling run_predict() is returned as "2.Predicted".

model_id : character. Type of fit model.

target : character. Name of target variable.

positive : character. Level of positive class of binary classification.

fitted_model : list. Fitted model object.

predicted : list. Predicted value by individual model. Each value has a predict_class class object.

Details

Supported models are functions supported by the representative model package used in R environment. The following binary classifications are supported:

"logistic" : logistic regression by predict.glm() in stats package.

"rpart" : recursive partitioning tree model by predict.rpart() in rpart package.

"ctree" : conditional inference tree model by predict() in stats package.

"randomForest" : random forest model by predict.randomForest() in randomForest package.

"ranger" : random forest model by predict.ranger() in ranger package.

run_predict() is executed in parallel when predicting by model. However, it is not supported in MS-Windows operating system and RStudio environment.

Examples

library(dplyr)

# Divide the train data set and the test data set.
sb <- rpart::kyphosis %>%
  split_by(Kyphosis)

# Extract the train data set from original data set.
train <- sb %>%
  extract_set(set = "train")

# Extract the test data set from original data set.
test <- sb %>%
  extract_set(set = "test")

# Sampling for unbalanced data set using SMOTE(synthetic minority over-sampling technique).
train <- sb %>%
  sampling_target(seed = 1234L, method = "ubSMOTE")

# Cleaning the set.
train <- train %>%
  cleanse

#> ── Checking unique value ─────────────────────────── unique value is one ──
#> No variables that unique value is one.
#> 
#> ── Checking unique rate ─────────────────────────────── high unique rate ──
#> No variables that high unique rate.
#> 
#> ── Checking character variables ─────────────────────── categorical data ──
#> No character variables.
#> 
#> 

# Run the model fitting.
result <- run_models(.data = train, target = "Kyphosis", positive = "present")
result

#> # A tibble: 5 x 5
#>   step     model_id     target   positive fitted_model
#>   <chr>    <chr>        <chr>    <chr>    <list>      
#> 1 1.Fitted logistic     Kyphosis present  <glm>       
#> 2 1.Fitted rpart        Kyphosis present  <rpart>     
#> 3 1.Fitted ctree        Kyphosis present  <BinaryTr>  
#> 4 1.Fitted randomForest Kyphosis present  <rndmFrs.>  
#> 5 1.Fitted ranger       Kyphosis present  <ranger>    

# Predict the model.
pred <- run_predict(result, test)
pred

#> # A tibble: 5 x 6
#>   step        model_id     target   positive fitted_model predicted 
#>   <chr>       <chr>        <chr>    <chr>    <list>       <list>    
#> 1 2.Predicted logistic     Kyphosis present  <glm>        <fct [24]>
#> 2 2.Predicted rpart        Kyphosis present  <rpart>      <fct [24]>
#> 3 2.Predicted ctree        Kyphosis present  <BinaryTr>   <fct [24]>
#> 4 2.Predicted randomForest Kyphosis present  <rndmFrs.>   <fct [24]>
#> 5 2.Predicted ranger       Kyphosis present  <ranger>     <fct [24]>

# Run the several kinds model predict by dplyr
result %>%
  run_predict(test)

#> # A tibble: 5 x 6
#>   step        model_id     target   positive fitted_model predicted 
#>   <chr>       <chr>        <chr>    <chr>    <list>       <list>    
#> 1 2.Predicted logistic     Kyphosis present  <glm>        <fct [24]>
#> 2 2.Predicted rpart        Kyphosis present  <rpart>      <fct [24]>
#> 3 2.Predicted ctree        Kyphosis present  <BinaryTr>   <fct [24]>
#> 4 2.Predicted randomForest Kyphosis present  <rndmFrs.>   <fct [24]>
#> 5 2.Predicted ranger       Kyphosis present  <ranger>     <fct [24]>